python_service/requirements.txt
flask==2.2.3
werkzeug==2.2.3
requests==2.26.0
python-dotenv==0.19.1

